fix: replace all SQLite BOOLEAN integer patterns with PostgreSQL bool types
Convert all hub-crate code that stored/read BOOLEAN columns as SQLite
integers (0/1) to use the PostgreSQL-native bool type throughout:
- Struct fields: is_bot, is_category, display_separately, is_pinned,
  is_locked, is_encrypted, is_group_encrypted, privileged, enabled,
  required, delivered, requires_camera changed from i64 to bool
- SQL literals: IS_CATEGORY=0/1 → false/true, is_bot=1 → TRUE, etc.
- Bindings: `if flag { 1i64 } else { 0 }` → `.bind(flag)`
- Query scalars: query_scalar::<_, i64> → query_scalar::<_, bool>
- Removed all `!= 0` and `== 1` comparisons on boolean fields

+++ crates/hub/src/bootstrap.rs
@@ -113,7 +113,7 @@ async fn apply_template(db: &PgPool, template: &serde_json::Value) -> anyhow::Re
             )
             .bind(&id)
             .bind(name)
-            .bind(is_category as i64)
+            .bind(is_category)
             .bind(now)
             .execute(db)
             .await
